% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2021.globalETtrend-package.R
\docType{package}
\name{p2021.globalETtrend-package}
\alias{p2021.globalETtrend}
\alias{p2021.globalETtrend-package}
\title{p2021.globalETtrend: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
